Added exception and status code errors handlers with their views to inform and guide the user when events like these happen.

+++ Web/Spotopedia.Web/Controllers/ChallengeEntriesController.cs
@@ -78,6 +78,13 @@ namespace Spotopedia.Web.Controllers
         [Authorize]
         public async Task<IActionResult> Delete(int id)
         {
+            var userId = this.userManager.GetUserId(this.User);
+
+            if (!this.challengeEntriesService.IsThisChallengeEntryAddedByThisUser(id, userId))
+            {
+                return this.RedirectToAction("StatusCodeForbidenError", "Home");
+            }
+
             await this.challengeEntriesService.DeleteAsync(id);
 
             this.TempData["DeleteEntry"] = $"Your entry was successfully deleted!";
